requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
websockets==11.0.3
python-dotenv==1.0.0
python-jose==3.3.0
pydantic==2.6.1
httpx==0.27.2
pytest
pytest-asyncio
sqlalchemy
psycopg2-binary
alembic
python-dotenv
